Tests: Add setup/teardown of TA files
Adds support to load a Trusted Application (TA) by filename before
starting a test suite and then remove the TA after the same test
suite has finished.

For every test suite, tee-supplicant is launched once the TA is in
place and pkill tee-supplicant is used after the test suite has
finished.

If there is an existing file with the same path, it is restored
after the test suite has ended. This is achieved by moving the
original file with the path %s to a new file with a path %s.bak,
then moving it back upon completion of the test-suite.

Change-Id: I3fc46d675cb3a7da1c293c743e77eab42917a645
Signed-off-by: Kurtis Charnock <kurtis.charnock@Arm.com>
Signed-off-by: Florian Depraz <florian.depraz@arm.com>