CMakeLists.txt: Integrate OP-TEE client
Finds libteec.so and the OP-TEE client header files.

If the library is installed, FindOPTEECLIENT.cmake will make use of
find_library and find_path to look for the necessary files.

If the library is not installed, the user can specify OPTEECLIENT_DIR to
the root of OP-TEE Client source directory.

Change-Id: I480ff9cfcb60fca5ca54d5f536916a3b64921db3
Signed-off-by: Kurtis Charnock <kurtis.charnock@Arm.com>
Signed-off-by: Florian Depraz <florian.depraz@arm.com>