libddssec: Replace DSEC_ARRAY_SIZE
DSEC_ARRAY_SIZE uses language extensions and, when used to find the size
of object handles, forces the arrays to be declared statically and the
TA to be recompiled every time they are updated.

This patch replaces DSEC_ARRAY_SIZE(x) with sizeof(x)/sizeof(x[0]) for
static arrays and with const variables for arrays which are allocated
dynamically or may be allocated dynamically in the future.

Issue-Id: SCM-694
Change-Id: I316fb89d24f898d502f757a8856fec14b908860e
Signed-off-by: Kurtis Charnock <kurtis.charnock@arm.com>