ta: Add key material module
Adds function to create, generate, copy, register and delete key materials.
Adds also a function to return the fields of a key material.
All those functions uses ID to retrieve the corresponding key material
stored in the TA.

Adds tests of the functionality for all the different functions.

Adds a note to the changelog to warn about the key material leak.

Change-Id: I983a7d72a914730d0c03fc8e35bf993719d00c58
Signed-off-by: Florian Depraz <florian.depraz@arm.com>