Update HMAC module to use mbedTLS
The GP API does not offer the possibility to have a 16 bytes key while
using HMAC256 (as written in the GPD_TEE_Internal_Core_API_Specification_v1.1):
The key should be between 192 and 1024 bits, multiple of 8. 16 bytes
(128 bits) is not supported.

Instead of using GP API, this patch changes to mbedTLS.

Change-Id: I3e81389fd4cf13529190491a2fe5721682e257e7
Signed-off-by: Florian Depraz <florian.depraz@arm.com>