Validation: Avoid re-building binaries in the target
This patch adds support for using the binaries from the build step in
validation.py instead of rebuilding the tests on the target. The
rationale for this change is to speed-up the validation as using
cross-compilation takes a fraction of the time when compared to a build
in the FastModels.

In order to achieve this, this patch removes the --prebuild-path
argument and add a new --build-on-target options. By default,
validation.py will always re-use the binaries from the build step. When
--build-on-target is passed to validation.py, them the test suites are
built on the target before executing them.

Change-Id: Ie8f492ff53edfa42a99e7f6bc202be9f49de233b
Signed-off-by: Filipe Rinaldi <filipe.rinaldi@arm.com>